requirements.txt
astroid==3.0.2
biopython==1.83
contourpy==1.0.7
cycler==0.11.0
dill==0.3.8
fonttools==4.39.4
isort==5.13.2
kiwisolver==1.4.4
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.26.3
packaging==23.1
pandas==2.2.0
Pillow==9.5.0
platformdirs==4.2.0
pyCirclize==1.2.0
pylint==3.0.3
pyparsing==3.0.9
pysam==0.22.0
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
six==1.16.0
tomlkit==0.12.3
tzdata==2023.4
